split module type declaration to own file
